experiments/beatles-and-you/requirements.txt
beautifulsoup4
requests
ratelimit
progressbar2
retry
